yolo/requirements.txt
ultralytics==8.3.162
mlflow==2.19.0
PyYAML==6.0.1
torch==2.0.1
opencv-python==4.8.0.76
numpy==1.26.1
